Setting AssemblyFileVersion to NugetVersion
Setting the AssemblyFileVersion to the NugetVersion string allows for
using the AssemblyFileVersion to check against Nuget to verify that you
are runnig the latest version of the assembly. The use case for this is
self-updating applications like nuget.exe. If Microsoft where to provide
custom assembly attribute (key, values) we could have used that.