Added link to ReleaseNotes.md for Choco Package
Since Chocolatey.org now supports markdown from nuspec file, adding in link to the full release notes stored in GitHub repo.  Decided not to add the full text for the release notes in the nuspec file, as this is just one more thing that needs to be remembered during the release.  Rather, point users to the one source of truth for release notes.