(#173) Fix up merge message for Stash
After actual testing on a live Stash server, it is clear that the message generated from Stash is of the format:

refs/heads/pull-requests/5/merge

and not as first thought:

refs/heads/pull-requests/5/merge-clean

Updated Acceptance Test, and also the Regex which identifies the messages, and tested on a Pull Request issued to server.  All seems to work as expected.