This changes AssemblyVersioningScheme.MajorMinorPatchMetadata to AssemblyVersioningScheme.MajorMinorPatchTag
The issue with metadata is that it is not part of the semantic version at all, so why are we even allowing it to be put into the assembly version. On the other hand in continuous delivery we promote metadata to the tag to allow the version to change every commit. By using the tag, not the metadata we can create meaningful AssemblySemVer values.

Fixes #391