Added branches in default config to GitVersion init branch list
GitVersion init branch configuration only includes branches which have custom config, this made it not very useful. Now at least you don't have to look at the source to se the branch regexes. We probably should take the next step of creating an alias so they key doesn't have to be a regex