Fix for issue #1162
This should fix the issue with GitVersionTask using the solution
directory .git folder instead of a project level .git folder, this is
more in line with how the command line works. It also adds property
support so the current use of the solution dir can be used or another
location can be specified. Also adds the ability to define a project
level or solution level import for global configuration.