Improved version calculation for feature branches in mainline mode - fixes #1154
In short when calculating the direct commits to master we need to remove any commits which are not on master. Otherwise we bump the mainline version on feature branches way too much. We also need to increment feature branches at the end of feature calculation unless the HEAD is a merge commit into the mainline branch (to cater for pull requests).