Fix GitVersionTask on Linux
The additional import statement introduced in #1163 causes failure
on Linux when a project tries to use the GitVersionTask. It seems
that Mono can't cope with an empty `Project` attribute in the
`Import` element. Additionally there seems to be a bug at least in
Mono 3.x in that it requires the argument in `Exists` to be passed
in quotes.

This change fixes these two problems and allows to use the
GitVersionTask again on Linux.