Ensure GetVersion runs prior to XAML's Markup Compiler
The XAML Markup Compiler (from PresentationBuildTasks via target
"MarkupCompilePass1") generates C# code which loads the XAML at run
time from a managed resource. This resource is identified by an URI,
containing the assembly where the resource is embedded and the path
towards it.
Without the "AssemblyVersion" property being set, this URI contains
only the simple assembly name, without version information. This
works as long as the assembly is only loaded once. However, in e.g.
advanced plugin scenarios it may happen that the same (strong named)
assembly is loaded several times in different versions. In those
cases the XAML resource can only be located when the version
information is available in the URI.

This fixes GitTools/GitVersion#1465.