Two additional heuristics for choosing mainline
When there is more than one potential mainline branch, before falling
back to the first branch in the list, choose 1) the current branch, if
it is a possible mainline; or 2) the first potential mainline from which
the current commit is reachable in a FirstParentOnly walk (i.e. is a
direct commit on that branch).

Additionally, don't discard potential mainlines with equal tip because
it might reduce the effectiveness of the new heuristics.

Fixes the test MainlineDevelopmentMode.VerifySupportForwardMerge vs the
previous commit.