(#2664) Add redirects for pages that have moved
As a result of changes to the structure of the documentation site, the
original 301 redirects from readthedocs, were resulting in 404's within
GitHub pages. This PR addresses this by ensuring that a Wyam
RedirectFrom attribute is added to all affected pages, so that the end
result is arrive on the correct page.

There may be additional changes that need to be made to capture all
pages, but this commit contains all 301 redirects from the current
readthedocs configuration.