IncrementStrategyFinder: Cache commit increments
Calculating version increments from commit messages is a relatively
expensive operation, involving up to four regex executions each.  As the
number of commits and tags in repositories grows, the number of
calculations performed grows at a super-linear rate, quickly adding
significantly to calculation run time.

This commit caches the results by commit sha, bringing this growth rate
back down to linear.

On a real-world repo with ~20,000 commits and ~5,000 tags, this halves
the GitVersion runtime from around 3 minutes down to around a minute and
a half.