IncrementStrategyFinder: Improve DetermineIncrementedField() performance
This optimizes a hot path under BaseVersionCalculator.GetBaseVersion(),
which among other things analyses commit messages on the sequence of
commits from candidate base versions to HEAD.  Candidate base versions
can include all version tags in the repository.

The previous implementation did an iteration of the full sequence of
commits from beginning to HEAD (which can number in the tens of
thousands or more) to get the commit subsequence for each of the
candidate base versions (which can number in the thousands, especially
in continuous deployment situations), so effectively a nested loop,
which quickly reached huge numbers of iterations as the number of
commits and tags in a repo increased.  Even though the commit message
analysis work itself was already being cached and therefore not
repeated, just the iteration itself could take significant amounts of
time.

The new implementation maintains a cache of the full sequence of commits
to HEAD (as the old one did, except as an array and with simpler code)
plus a lookup indexing into that sequence by commit sha.  To get the
subsequence for each base version, we get its position in the full
sequence via the lookup and yield a zero-allocation ArraySegment from
there to the end.  This eliminates the nested loop, reducing the whole
thing to essentially a linear iteration over the base versions (tags).

On a real-world repo with ~20,000 commits and ~5,000 tags, this brings
the run time down from around 30 seconds to around 20 seconds.