Extract GitVersion Configuration into a separate project
The major change in this commit is the extraction of the GitVersion configuration-related code into a separate project. This includes branch configuration, GitFlow and GitHub workflows, caching strategies and modules. In the refactored structure, other projects now reference this configuration project. The definition of the `IBranchConfiguration` has changed and includes a new method for providing an empty configuration. Some necessary code and file shifting has occurred to support this new structure.